isolando types
