+++ src/models/user.ts
@@ -1,26 +1,9 @@
-import { DataTypes, Model, BuildOptions } from "sequelize";
+import { DataTypes } from "sequelize";
 import db from "../configs/db";
 import bcrypt from "bcryptjs";
-import { PlanModel } from "./plan";
+import { UserStatic, TAuthentication } from "../types/models/user";
 
-type TAuthentication = Promise<UserModel | null>;
 
-interface UserAttributes {
-    name: string;
-    email: string;
-    password: string;
-    type: string;
-}
-
-export interface UserModel extends Model<UserAttributes>, UserAttributes {
-    id: number;
-    getPlans(): PlanModel[];
-    addPlan(plan: PlanModel): null;
-};
-type UserStatic = typeof Model & {
-    new (values?: object, options?: BuildOptions): UserModel;
-    authenticate(email: string, password: string): TAuthentication;
-};
 
 const User = <UserStatic>db.define("user", {
     name: {
